examples/facr-chatbot/backend/pyproject.toml
[project]
name = "facr-chatbot-backend"
version = "0.1.0"
description = "FastAPI backend for FAČR knowledge-base chatbot powered by ChatKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-chatkit>=1.6.0,<1.7.0",
    "numpy>=1.26,<3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]
